requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Markdown==3.8.2
pdfkit==1.0.0
pillow==11.3.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
requests==2.32.5
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
